Redundant casts

+++ tryonline/src/main/java/KillProcessActionProperty.java
@@ -1,14 +1,12 @@
 import com.google.common.base.Throwables;
-import lsfusion.server.physics.dev.integration.internal.to.InternalAction;
-import lsfusion.server.logics.classes.ValueClass;
-import lsfusion.server.data.sql.exception.SQLHandledException;
 import lsfusion.server.data.value.DataObject;
-import lsfusion.server.logics.property.classes.ClassPropertyInterface;
-import lsfusion.server.logics.action.controller.context.ExecutionContext;
 import lsfusion.server.language.ScriptingLogicsModule;
+import lsfusion.server.logics.action.controller.context.ExecutionContext;
+import lsfusion.server.logics.classes.ValueClass;
+import lsfusion.server.logics.property.classes.ClassPropertyInterface;
+import lsfusion.server.physics.dev.integration.internal.to.InternalAction;
 
 import java.lang.ref.WeakReference;
-import java.sql.SQLException;
 import java.util.Iterator;
 
 public class KillProcessActionProperty extends InternalAction {
@@ -26,7 +24,7 @@ public class KillProcessActionProperty extends InternalAction {
         DataObject server = (DataObject) context.getKeyValue(serverInterface);
 
         try {
-            WeakReference<Process> weakProcess = RunProcessActionProperty.runningProcesses.get((Long) server.object);
+            WeakReference<Process> weakProcess = RunProcessActionProperty.runningProcesses.get(server.object);
             Process p;
             if(weakProcess != null && (p = weakProcess.get()) != null) {
                 p.destroy();
